phase-5-till-api-extraction/requirements.txt
# requirements.txt
typer>=0.6.1
rich>=12.0.0
pydantic>=1.9.0
duckdb>=0.7.0
requests>=2.27.1
spacy>=3.5.0
pyyaml
beautifulsoup4
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart
jinja2>=3.1.2
aiofiles
python-dotenv>=1.0.0